Move Node/LockedNode/LockFileV7/LockedFlakeV7 into lockfile-v7.cc
These types are no longer used outside of lockfile-v7.cc, so they can
be implementation details. The header now only exposes two free
functions returning `std::unique_ptr<LockedFlake>`:

* `parseLockFileV7()`: construct a `LockedFlake` from the JSON
  contents of a version 5-7 lock file.

* `lockFlakeV7()` (previously `LockedFlakeV7::lockFlake()`): compute a
  new lock file.

This also keeps the version dispatch in the free function `lockFlake()`
symmetrical for the future version 8 implementation: it will just pick
between `parseLockFileV{7,8}()` and `lockFlakeV{7,8}()` based on the
`version` field in the lock file JSON.

Also remove the unused `LockedNode::computeStorePath()`.
