+++ src/libflake/lockfile-v7.cc
@@ -22,6 +22,7 @@
 
 #include "nix/fetchers/fetch-settings.hh"
 #include "nix/flake/lockfile-v7.hh"
+#include "nix/util/sync.hh"
 #include "nix/flake/settings.hh"
 #include "nix/expr/eval.hh"
 #include "nix/expr/nixexpr.hh"
@@ -48,6 +49,129 @@ class Store;
 } // namespace nix
 
 namespace nix::flake {
+
+struct LockedNode;
+
+/**
+ * A node in the lock file. It has outgoing edges to other nodes (its
+ * inputs). Only the root node has this type; all other nodes have
+ * type LockedNode.
+ */
+struct Node : std::enable_shared_from_this<Node>
+{
+    typedef std::variant<ref<LockedNode>, InputAttrPath> Edge;
+
+    std::map<FlakeId, Edge> inputs;
+
+    virtual ~Node() {}
+};
+
+/**
+ * A non-root node in the lock file.
+ */
+struct LockedNode : Node
+{
+    FlakeRef lockedRef, originalRef;
+    bool isFlake = true;
+    bool buildTime = false;
+
+    /* The node relative to which relative source paths
+       (e.g. 'path:../foo') are interpreted. */
+    std::optional<InputAttrPath> parentInputAttrPath;
+
+    /**
+     * The source path of this node, if it has been fetched. Set by
+     * `LockedFlakeV7::lockFlake()` for nodes fetched during locking,
+     * and by `LockedFlakeV7::getSourcePath()` for nodes fetched on
+     * demand.
+     */
+    mutable Sync<std::optional<SourcePath>> sourcePath;
+
+    LockedNode(
+        const FlakeRef & lockedRef,
+        const FlakeRef & originalRef,
+        bool isFlake = true,
+        bool buildTime = false,
+        std::optional<InputAttrPath> parentInputAttrPath = {})
+        : lockedRef(std::move(lockedRef))
+        , originalRef(std::move(originalRef))
+        , isFlake(isFlake)
+        , buildTime(buildTime)
+        , parentInputAttrPath(std::move(parentInputAttrPath))
+    {
+    }
+
+    LockedNode(const fetchers::Settings & fetchSettings, const nlohmann::json & json);
+};
+
+/**
+ * The old graph-based lock file format (versions 5-7).
+ */
+struct LockFileV7
+{
+    ref<Node> root = make_ref<Node>();
+
+    LockFileV7() {};
+    LockFileV7(const fetchers::Settings & fetchSettings, const nlohmann::json & json, std::string_view path);
+
+    nlohmann::json toJSON() const;
+
+    /**
+     * Check whether this lock file has any unlocked or non-final
+     * inputs. If so, return one.
+     */
+    std::optional<FlakeRef> isUnlocked(const fetchers::Settings & fetchSettings) const;
+
+    bool operator==(const LockFileV7 & other) const;
+
+    std::shared_ptr<Node> findInput(const InputAttrPath & path) const;
+
+    std::map<InputAttrPath, Node::Edge> getAllInputs() const;
+
+    /**
+     * Check that every 'follows' input target exists.
+     */
+    void check();
+};
+
+struct LockedFlakeV7 : LockedFlake
+{
+    /**
+     * The lock file in the old graph-based format (versions 5-7).
+     */
+    LockFileV7 lockFile;
+
+    LockedFlakeV7(Flake && flake, LockFileV7 && lockFile)
+        : LockedFlake(std::move(flake))
+        , lockFile(std::move(lockFile))
+    {
+    }
+
+    /**
+     * Construct from the JSON contents of a lock file (which must be
+     * null if the lock file doesn't exist).
+     */
+    LockedFlakeV7(
+        const fetchers::Settings & fetchSettings, Flake flake, const nlohmann::json & json, std::string_view path)
+        : LockedFlake(std::move(flake))
+        , lockFile(json.is_null() ? LockFileV7() : LockFileV7(fetchSettings, json, path))
+    {
+    }
+
+    std::map<FlakeId, std::optional<InputAttrPath>> getInputTargets(const InputAttrPath & prefix) const override;
+
+    std::optional<InputInfo> findInput(const InputAttrPath & path) const override;
+
+    SourcePath getSourcePath(EvalState & state, const InputAttrPath & inputAttrPath) const override;
+
+    void visit(VisitCallback callback) const override;
+
+    std::optional<FlakeRef> isUnlocked(const fetchers::Settings & fetchSettings) const override;
+
+    std::string diff(const LockedFlake & oldLockFile) const override;
+
+    nlohmann::json toJSON() const override;
+};
 
 static FlakeRef
 getFlakeRef(const fetchers::Settings & fetchSettings, const nlohmann::json & json, const char * attr, const char * info)
@@ -92,11 +216,6 @@ LockedNode::LockedNode(const fetchers::Settings & fetchSettings, const nlohmann:
     // For backward compatibility, lock file entries are implicitly final.
     assert(!lockedRef.input.attrs.contains("__final"));
     lockedRef.input.attrs.insert_or_assign("__final", Explicit<bool>(true));
-}
-
-StorePath LockedNode::computeStorePath(Store & store) const
-{
-    return lockedRef.input.computeStorePath(store);
 }
 
 static std::shared_ptr<Node>
@@ -500,13 +619,6 @@ nlohmann::json LockedFlakeV7::toJSON() const
     return lockFile.toJSON();
 }
 
-LockedFlakeV7::LockedFlakeV7(
-    const fetchers::Settings & fetchSettings, Flake flake, const nlohmann::json & json, std::string_view path)
-    : LockedFlake(std::move(flake))
-    , lockFile(json.is_null() ? LockFileV7() : LockFileV7(fetchSettings, json, path))
-{
-}
-
 static LockFileV7 readLockFile(const fetchers::Settings & fetchSettings, const SourcePath & lockFilePath)
 {
     if (!lockFilePath.pathExists())
@@ -523,7 +635,13 @@ static LockFileV7 readLockFile(const fetchers::Settings & fetchSettings, const S
     return LockFileV7(fetchSettings, json, fmt("%s", lockFilePath));
 }
 
-std::unique_ptr<LockedFlake> LockedFlakeV7::lockFlake(
+std::unique_ptr<LockedFlake> parseLockFileV7(
+    const fetchers::Settings & fetchSettings, Flake flake, const nlohmann::json & json, std::string_view path)
+{
+    return std::make_unique<LockedFlakeV7>(fetchSettings, std::move(flake), json, path);
+}
+
+std::unique_ptr<LockedFlake> lockFlakeV7(
     const Settings & settings,
     EvalState & state,
     const LockFlags & lockFlags,
